+++ SQLiteStudio3/guiSQLiteStudio/uiconfig.h
@@ -130,6 +130,7 @@ CFG_UI_CATEGORIES(Ui,
         CFG_ENTRY(bool,                    SqlEditorWrapWords,          false)
         CFG_ENTRY(bool,                    SqlEditorCurrQueryHighlight, true)
         CFG_ENTRY(qint64,                  SqlEditorExecTimePrecision,  6)
+        CFG_ENTRY(bool,                    SqlEditorSnippetsIfEmpty,    true)
         CFG_ENTRY(bool,                    DisableBlinkingCursor,       false)
         CFG_ENTRY(bool,                    ExecuteCurrentQueryOnly,     true)
         CFG_ENTRY(int,                     NumberOfRowsPerPage,         1000)